+++ changelog.md
@@ -1,2 +1,2 @@
 ### SuperMartijn642's Core Library 1.1.18
-- Initial release of SuperMartijn642's Core Library for Minecraft 1.21.6
+- Initial release of SuperMartijn642's Core Library for Minecraft 1.21.7